MLM-EOE-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.5.0=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
args=0.1.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
audioread=3.0.1=pypi_0
axial-positional-embedding=0.3.12=pypi_0
ca-certificates=2025.2.25=h06a4308_0
causal-conv1d=1.2.2.post1=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
clint=0.5.1=pypi_0
clip-openai=1.0.post20230121=pypi_0
cmake=3.31.6=pypi_0
colt5-attention=0.11.1=pypi_0
contourpy=1.3.0=pypi_0
coverage=7.6.12=pypi_0
cycler=0.12.1=pypi_0
decorator=5.2.1=pypi_0
einops=0.8.1=pypi_0
et-xmlfile=2.0.0=pypi_0
filelock=3.17.0=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2025.3.0=pypi_0
ftfy=6.3.1=pypi_0
huggingface-hub=0.29.2=pypi_0
hyper-connections=0.1.15=pypi_0
idna=3.10=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.11.0=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
lightning=2.5.0.post0=pypi_0
lightning-utilities=0.14.0=pypi_0
llvmlite=0.43.0=pypi_0
local-attention=1.11.1=pypi_0
mamba-ssm=1.2.2=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.1=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
ninja=1.11.1.3=pypi_0
nnaudio=0.3.3=pypi_0
numba=0.60.0=pypi_0
numpy=2.0.2=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.11.0.86=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
patsy=1.0.1=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py39h06a4308_0
platformdirs=4.3.8=pypi_0
pooch=1.8.2=pypi_0
product-key-memory=0.2.11=pypi_0
propcache=0.3.0=pypi_0
psutil=7.0.0=pypi_0
pycparser=2.22=pypi_0
pyparsing=3.2.1=pypi_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.5.0.post0=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
reformer-pytorch=1.4.4=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.8.0=py39h06a4308_0
six=1.17.0=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.45.3=h5eee18b_0
statsmodels=0.14.4=pypi_0
sympy=1.14.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=0.4.5=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
torch=2.4.0=pypi_0
torchaudio=2.4.0=pypi_0
torchlibrosa=0.1.0=pypi_0
torchlightning=0.0.0=pypi_0
torchmetrics=1.6.2=pypi_0
torchvision=0.19.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.49.0=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.45.1=py39h06a4308_0
xz=5.6.4=h5eee18b_1
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
